Renamed errors.py to _exceptions.py
The BoltError are internal errors that the user should never experience.
